FEAT-#2197: move to_datetime implementation into general.py
