--- modin/pandas/datetimes.py
@@ -1,140 +1,0 @@
-# Licensed to Modin Development Team under one or more contributor license agreements.
-# See the NOTICE file distributed with this work for additional information regarding
-# copyright ownership.  The Modin Development Team licenses this file to you under the
-# Apache License, Version 2.0 (the "License"); you may not use this file except in
-# compliance with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific language
-# governing permissions and limitations under the License.
-
-import pandas
-
-from .dataframe import DataFrame
-from .series import Series
-
-
-def to_datetime(
-    arg,
-    errors="raise",
-    dayfirst=False,
-    yearfirst=False,
-    utc=None,
-    format=None,
-    exact=True,
-    unit=None,
-    infer_datetime_format=False,
-    origin="unix",
-    cache=True,
-):
-    """
-    Convert argument to datetime.
-
-    Parameters
-    ----------
-    arg : int, float, str, datetime, list, tuple, 1-d array, Series, DataFrame/dict-like
-        The object to convert to a datetime.
-    errors : {'ignore', 'raise', 'coerce'}, default 'raise'
-        - If 'raise', then invalid parsing will raise an exception.
-        - If 'coerce', then invalid parsing will be set as NaT.
-        - If 'ignore', then invalid parsing will return the input.
-    dayfirst : bool, default False
-        Specify a date parse order if `arg` is str or its list-likes.
-        If True, parses dates with the day first, eg 10/11/12 is parsed as
-        2012-11-10.
-        Warning: dayfirst=True is not strict, but will prefer to parse
-        with day first (this is a known bug, based on dateutil behavior).
-    yearfirst : bool, default False
-        Specify a date parse order if `arg` is str or its list-likes.
-
-        - If True parses dates with the year first, eg 10/11/12 is parsed as
-          2010-11-12.
-        - If both dayfirst and yearfirst are True, yearfirst is preceded (same
-          as dateutil).
-
-        Warning: yearfirst=True is not strict, but will prefer to parse
-        with year first (this is a known bug, based on dateutil behavior).
-    utc : bool, default None
-        Return UTC DatetimeIndex if True (converting any tz-aware
-        datetime.datetime objects as well).
-    format : str, default None
-        The strftime to parse time, eg "%d/%m/%Y", note that "%f" will parse
-        all the way up to nanoseconds.
-        See strftime documentation for more information on choices:
-        https://docs.python.org/3/library/datetime.html#strftime-and-strptime-behavior.
-    exact : bool, True by default
-        Behaves as:
-        - If True, require an exact format match.
-        - If False, allow the format to match anywhere in the target string.
-
-    unit : str, default 'ns'
-        The unit of the arg (D,s,ms,us,ns) denote the unit, which is an
-        integer or float number. This will be based off the origin.
-        Example, with unit='ms' and origin='unix' (the default), this
-        would calculate the number of milliseconds to the unix epoch start.
-    infer_datetime_format : bool, default False
-        If True and no `format` is given, attempt to infer the format of the
-        datetime strings based on the first non-NaN element,
-        and if it can be inferred, switch to a faster method of parsing them.
-        In some cases this can increase the parsing speed by ~5-10x.
-    origin : scalar, default 'unix'
-        Define the reference date. The numeric values would be parsed as number
-        of units (defined by `unit`) since this reference date.
-
-        - If 'unix' (or POSIX) time; origin is set to 1970-01-01.
-        - If 'julian', unit must be 'D', and origin is set to beginning of
-          Julian Calendar. Julian day number 0 is assigned to the day starting
-          at noon on January 1, 4713 BC.
-        - If Timestamp convertible, origin is set to Timestamp identified by
-          origin.
-    cache : bool, default True
-        If True, use a cache of unique, converted dates to apply the datetime
-        conversion. May produce significant speed-up when parsing duplicate
-        date strings, especially ones with timezone offsets. The cache is only
-        used when there are at least 50 values. The presence of out-of-bounds
-        values will render the cache unusable and may slow down parsing.
-
-    Returns
-    -------
-    datetime
-        If parsing succeeded.
-        Return type depends on input:
-
-        - list-like: DatetimeIndex
-        - Series: Series of datetime64 dtype
-        - scalar: Timestamp
-
-        In case when it is not possible to return designated types (e.g. when
-        any element of input is before Timestamp.min or after Timestamp.max)
-        return will have datetime.datetime type (or corresponding
-        array/Series).
-    """
-    if not isinstance(arg, (DataFrame, Series)):
-        return pandas.to_datetime(
-            arg,
-            errors=errors,
-            dayfirst=dayfirst,
-            yearfirst=yearfirst,
-            utc=utc,
-            format=format,
-            exact=exact,
-            unit=unit,
-            infer_datetime_format=infer_datetime_format,
-            origin=origin,
-            cache=cache,
-        )
-    return arg._to_datetime(
-        errors=errors,
-        dayfirst=dayfirst,
-        yearfirst=yearfirst,
-        utc=utc,
-        format=format,
-        exact=exact,
-        unit=unit,
-        infer_datetime_format=infer_datetime_format,
-        origin=origin,
-        cache=cache,
-    )
